backend/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
torch
torchvision
transformers
opencv-python-headless
numpy
pillow
matplotlib
python-dotenv
spacy
thefuzz
python-Levenshtein
sqlalchemy
psycopg2-binary
timm
scikit-image
pdf2image
PyPDF2
grad-cam
